requirements.txt
annotated-types==0.7.0
anyio==4.6.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
distro==1.9.0
google-api-core==2.20.0
google-api-python-client==2.147.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.65.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.2
idna==3.10
jiter==0.5.0
openai==1.48.0
proto-plus==1.24.0
protobuf==5.28.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
pyparsing==3.1.4
python-dotenv==1.0.1
requests==2.32.3
rsa==4.9
sniffio==1.3.1
tqdm==4.66.5
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
youtube-transcript-api==0.6.2
